+++ src/module.js
@@ -1,26 +1,20 @@
 Fossil.Module = (function (Fossil, _, Backbone) {
     'use strict';
 
-    var Module = function (application, path, options) {
-        if (typeof path === "string") {
-            this.path = path;
-            this.options = options || {};
-        } else {
-            options = path || {};
-            this.path = options.path || '';
-            this.options = options;
+    var Module = function (options) {
+        this.options = options || {};
+        if (typeof this.options.path === "string") {
+            this.path = _.result(this.options, 'path');
         }
 
-        // a PubSub object for communication with the application
-        this.application = application.createPubSub(this, 'applicationEvents');
-        // init services namespace
-        this.services = {};
         // init event listeners
-        this.registerEvents(application);
+        this.registerEvents();
+        // init layoutable
+        this.initLayoutable();
         // init fragmentable
         this.initFragmentable();
         // finally call initialize method
-        this.initialize.call(this, application);
+        this.initialize.call(this);
     };
 
     _.extend(Module.prototype,
@@ -34,11 +28,21 @@ Fossil.Module = (function (Fossil, _, Backbone) {
             applicationEvents: {},
             // events bound on module PubSub
             events: {},
-            initialize: function (application) {
-
-            },
-            registerEvents: function (application) {
+            initialize: function () {},
+            registerEvents: function () {
                 Fossil.Mixins.Observable.registerEvents.call(this);
+                this.listenTo(this, 'connect', _.bind(this.connectListener, this));
+            },
+            connectListener: function (application, id) {
+                // a PubSub object to communicate with the application
+                this.application = application.createPubSub(this, 'applicationEvents');
+                // if not already defined
+                if (typeof this.path !== "string") {
+                    this.path = id;
+                }
+                // link services
+                this.services = application.services;
+                // start and stop when element is set or unset
                 this.listenTo(this, 'elementable:attach', _.bind(this.elementAttachListener, this, application));
                 this.listenTo(this, 'elementable:detach', _.bind(this.elementDetachListener, this, application));
             },
@@ -49,7 +53,7 @@ Fossil.Module = (function (Fossil, _, Backbone) {
             elementDetachListener: function (application) {
                 this.standby();
             },
-            render: function (application) {
+            render: function () {
                 this.renderLayout();
                 this.renderFragments();
             },